New info about usage as a Middleman Template
Previously, I thought that it would be best if the project was set up to be forked and then built on from there, but I now realize that that's not really a good use case at all.  It turns out that it could be used out of the box as a Middleman template, though, so I added instructions for doing that instead!